project-tee-starter: move __tests__ (component tests) inside of src folder
