+++ packages/project-tee-starter/src/__tests__/models.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'bun:test';
-import teeStarterPlugin from '../src/plugin';
+import teeStarterPlugin from '../plugin';
 
 describe('Plugin Models', () => {
   it('should not have custom models (using plugin-tee and other plugins for model handling)', () => {
